+++ modules/alloy/src/main/ts/ephox/alloy/dragging/common/DragMovement.ts
@@ -1,12 +1,13 @@
 import { Option } from '@ephox/katamari';
 import { Css, Location, Scroll, Traverse, Element } from '@ephox/sugar';
 
+import { cap } from '../../alien/Cycles';
 import * as OffsetOrigin from '../../alien/OffsetOrigin';
 import { SugarPosition } from '../../alien/TypeDefinitions';
 import { AlloyComponent } from '../../api/component/ComponentApi';
 import * as DragCoord from '../../api/data/DragCoord';
 import * as Snappables from '../snap/Snappables';
-import { DraggingConfig, SnapsConfig } from './DraggingTypes';
+import { DraggingConfig, DragStartData, SnapsConfig } from './DraggingTypes';
 
 const getCurrentCoord = (target: Element): DragCoord.CoordAdt => {
   return Css.getRaw(target, 'left').bind((left) => {
@@ -25,8 +26,33 @@ const getCurrentCoord = (target: Element): DragCoord.CoordAdt => {
   });
 };
 
-const calcNewCoord = (component: AlloyComponent, optSnaps: Option<SnapsConfig>, currentCoord: DragCoord.CoordAdt, scroll: SugarPosition, origin: SugarPosition, delta: SugarPosition): DragCoord.CoordAdt => {
-  return optSnaps.fold(() => {
+const clampCoords = (component: AlloyComponent, coords: DragCoord.CoordAdt, scroll: SugarPosition, origin: SugarPosition, startData: DragStartData): DragCoord.CoordAdt => {
+  const bounds = startData.bounds;
+  const absoluteCoord = DragCoord.asAbsolute(coords, scroll, origin);
+
+  const newX = cap(absoluteCoord.left(), bounds.x(), bounds.width() - startData.width);
+  const newY = cap(absoluteCoord.top(), bounds.y(), bounds.height() - startData.height);
+  const newCoords = DragCoord.absolute(newX, newY);
+
+  // Translate the absolute coord back into the previous type
+  return coords.fold(
+    // offset
+    () => {
+      const offset = DragCoord.asOffset(newCoords, scroll, origin);
+      return DragCoord.offset(offset.left(), offset.top());
+    },
+    // absolute
+    () => newCoords,
+    // fixed
+    () => {
+      const fixed = DragCoord.asFixed(newCoords, scroll, origin);
+      return DragCoord.fixed(fixed.left(), fixed.top());
+    },
+  );
+};
+
+const calcNewCoord = (component: AlloyComponent, optSnaps: Option<SnapsConfig>, currentCoord: DragCoord.CoordAdt, scroll: SugarPosition, origin: SugarPosition, delta: SugarPosition, startData: DragStartData): DragCoord.CoordAdt => {
+  const newCoord = optSnaps.fold(() => {
     // When not docking, use fixed coordinates.
     const translated = DragCoord.translate(currentCoord, delta.left(), delta.top());
     const fixedCoord = DragCoord.asFixed(translated, scroll, origin);
@@ -38,9 +64,12 @@ const calcNewCoord = (component: AlloyComponent, optSnaps: Option<SnapsConfig>, 
     });
     return snapping.coord;
   });
+
+  // Clamp the coords so that they are within the bounds
+  return clampCoords(component, newCoord, scroll, origin, startData);
 };
 
-const dragBy = (component: AlloyComponent, dragConfig: DraggingConfig, delta: SugarPosition): void => {
+const dragBy = (component: AlloyComponent, dragConfig: DraggingConfig, startData: DragStartData, delta: SugarPosition): void => {
   const target = dragConfig.getTarget(component.element());
 
   if (dragConfig.repositionTarget) {
@@ -51,7 +80,8 @@ const dragBy = (component: AlloyComponent, dragConfig: DraggingConfig, delta: Su
 
     const currentCoord = getCurrentCoord(target);
 
-    const newCoord = calcNewCoord(component, dragConfig.snaps, currentCoord, scroll, origin, delta);
+    const newCoord = calcNewCoord(component, dragConfig.snaps, currentCoord, scroll, origin, delta, startData);
+
     const styles = DragCoord.toStyles(newCoord, scroll, origin);
     Css.setAll(target, styles);
   }